+++ tests/test_job.py
@@ -21,8 +21,23 @@
 
 import os
 import stat
+from pathlib import Path
 
 from parman.job import JobFactory
+
+
+def setup_jobfactory(root: Path, jobinfo: str, run: str) -> JobFactory:
+    template_path = root / "templates" / "sample"
+    template_path.mkdir(parents=True)
+    with open(template_path / "jobinfo.py", "w") as f:
+        f.write(jobinfo)
+    with open(template_path / "run", "w") as f:
+        f.write(run)
+    os.chmod(template_path / "run", stat.S_IRWXU)
+    job = JobFactory()
+    job.clerk.root = root / "results"
+    return job, template_path
+
 
 ENVVAR_JOBINFO = """
 def get_result_mock() -> str:
@@ -46,17 +61,41 @@ if __name__ == "__main__":
 """
 
 
-def test_env_pass_through(tmppath):
-    sample_path = tmppath / "templates" / "sample"
-    sample_path.mkdir(parents=True)
-    with open(sample_path / "jobinfo.py", "w") as f:
-        f.write(ENVVAR_JOBINFO)
-    with open(sample_path / "run", "w") as f:
-        f.write(ENVVAR_RUN)
-    os.chmod(sample_path / "run", stat.S_IRWXU)
-    job = JobFactory()
-    job.clerk.root = tmppath / "results"
-    closure = job(sample_path, "sample")
+def test_envvar(tmppath: Path):
+    job, template_path = setup_jobfactory(tmppath, ENVVAR_JOBINFO, ENVVAR_RUN)
+    closure = job(template_path, "sample")
     os.putenv("PARMAN_TEST_JOB_QOFDHEIC", "correct")
     result = closure.validated_call()
     assert result == "correct"
+
+
+OPTIONAL_JOBINFO = """
+def get_result_mock(first: int, second: int = 2) -> int:
+    return 42
+"""
+
+OPTIONAL_RUN = """\
+#!/usr/bin/env python
+
+import json
+import os
+
+
+def main():
+    with open("kwargs.json") as f:
+        kwargs = json.load(f)
+
+    with open("result.json", "w") as f:
+        json.dump(kwargs["first"] + kwargs["second"], f)
+
+
+if __name__ == "__main__":
+    main()
+"""
+
+
+def test_optional(tmppath: Path):
+    job, template_path = setup_jobfactory(tmppath, OPTIONAL_JOBINFO, OPTIONAL_RUN)
+    closure = job(template_path, "sample", first=1)
+    result = closure.validated_call()
+    assert result == 3