+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - A pure Python usage example (Differential Evolution Monte Carlo).
 - The `parman-sbatch-wait` script to facilitate running workflow jobs on a Slurm cluster.
+- Support for optional arguments in `jobinfo.py`
 
 ### Fixed